iio: light: vcnl4000 add support for the VCNL4040 proximity and light sensor
The VCNL4040 is almost identical to the VCNL4200 as far as register
layout goes but just need to check a different ID register location.

This does change the initialization sequence of the VCNL4200 to use word writes
instead of byte writes. The VCNL4200 says that word read and writes should be
used to access the registers but I don't have a 4200 to test with. The VCNL4040
doesn't initialize properly with the byte writes.

Signed-off-by: Angus Ainslie (Purism) <angus@akkea.ca>